refactor when the search index is built so that all assets are processed first
this also moves the index over to utils.
